requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
python-dotenv>=1.0.0
openai>=1.0.0
plotly>=5.15.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-google-genai>=0.0.6
langchain-together>=0.0.5
langgraph>=0.0.20
tqdm>=4.64.0
httpx>=0.25.0
psycopg2-binary>=2.9.0
openpyxl>=3.1.0
typing-extensions>=4.8.0
langsmith>=0.0.67 